+++ src/Library/PhotonMapping/CausticSpectralPhotonMap.cpp
@@ -44,7 +44,7 @@ void CausticSpectralPhotonMap::RadianceEstimate(
 
 	// locate the nearest photons
 	PhotonDistListType heap;
-	LocatePhotons( ri.ptIntersection, dGatherRadius, nMaxPhotonsOnGather, heap, 0, vphotons.size()-1 );
+	LocatePhotons( ri.ptIntersection, dGatherRadius, nMaxPhotonsOnGather, heap, 0, static_cast<int>(vphotons.size())-1 );
 
 	if( heap.size() > nMinPhotonsOnGather )
 	{
@@ -100,7 +100,7 @@ void CausticSpectralPhotonMap::RadianceEstimateNM(
 
 	// locate the nearest photons
 	PhotonDistListType heap;
-	LocatePhotons( ri.ptIntersection, dGatherRadius, nMaxPhotonsOnGather, heap, 0, vphotons.size()-1 );
+	LocatePhotons( ri.ptIntersection, dGatherRadius, nMaxPhotonsOnGather, heap, 0, static_cast<int>(vphotons.size())-1 );
 
 	if( heap.size() > nMinPhotonsOnGather )
 	{
@@ -189,8 +189,8 @@ void CausticSpectralPhotonMap::Serialize(
 	bbox.Serialize( buffer );
 
 	// Serialize number of stored photons
-	buffer.ResizeForMore( sizeof( unsigned int ) + sizeof( SpectralPhoton ) * vphotons.size() );
-	buffer.setUInt( vphotons.size() );
+	buffer.ResizeForMore( static_cast<unsigned int>(sizeof( unsigned int ) + sizeof( SpectralPhoton ) * vphotons.size()) );
+	buffer.setUInt( static_cast<unsigned int>(vphotons.size()) );
 
 	for( unsigned int i=0; i<vphotons.size(); i++ ) {
 		const SpectralPhoton& p = vphotons[i];
