+++ src/Library/RasterImages/PPMWriter.cpp
@@ -51,10 +51,10 @@ void PPMWriter::BeginWrite( const unsigned int width, const unsigned int height 
 	snprintf( header, MAX_BUFFER_SIZE, "P6\n%d %d\n255\n", width, height );
 
 	// Make sure the buffer is the correct size
-	pWriteBuffer.Resize( strlen( header ) + width*height*3 );
+	pWriteBuffer.Resize( static_cast<unsigned int>(strlen( header ) + width*height*3) );
 
 	// Copy the header
-	pWriteBuffer.setBytes( header, strlen( header ) );
+	pWriteBuffer.setBytes( header, static_cast<unsigned int>(strlen( header )) );
 	
 	bufW = width;
 	bufH = height;
