+++ src/Library/RasterImages/RGBEAWriter.cpp
@@ -60,10 +60,10 @@ void RGBEAWriter::BeginWrite( const unsigned int width, const unsigned int heigh
 
 	// Write the header
 	// Make sure there's enough room in the buffer first
-	pWriteBuffer.Resize( width*height*8 + strlen(szHeader) );
+	pWriteBuffer.Resize( static_cast<unsigned int>(width*height*8 + strlen(szHeader)) );
 
 	// Write out the header
-	pWriteBuffer.setBytes( szHeader, strlen(szHeader) );
+	pWriteBuffer.setBytes( szHeader, static_cast<unsigned int>(strlen(szHeader)) );
 
 	// Setup the buffer
 	bufW = width;
